+++ src/flavors/cdc.js
@@ -1,9 +1,12 @@
 import {
   printStartPipeline, printEndPipeline,
   faulty, faultyAsyncStream, faultify,
-  queryDynamoDB, batchGetDynamoDB, toPkQueryRequest,
   encryptEvent, compact,
 } from '../utils';
+
+import {
+  queryAllDynamoDB, batchGetDynamoDB, toPkQueryRequest,
+} from '../queries/dynamodb';
 
 import { filterOnEventType, filterOnContent, outLatched } from '../filters';
 
@@ -18,7 +21,7 @@ export const cdc = (rule) => (s) => s // eslint-disable-line import/prefer-defau
   .through(compact(rule))
 
   .map(toQueryRequest(rule))
-  .through(queryDynamoDB(rule))
+  .through(queryAllDynamoDB(rule))
 
   .map(toGetRequest(rule))
   .through(batchGetDynamoDB(rule))
