+++ src/queries/dynamodb.js
@@ -1,110 +1,10 @@
 import _ from 'highland';
-import merge from 'lodash/merge';
 import memoryCache from 'memory-cache';
 
 import Connector from '../connectors/dynamodb';
 
-import { rejectWithFault } from './faults';
-import { debug as d } from './print';
-import { ratelimit } from './ratelimit';
-
-export const updateExpression = (Item) => {
-  const keys = Object.keys(Item);
-
-  const ExpressionAttributeNames = keys
-    .filter((attrName) => Item[attrName] !== undefined)
-    .map((attrName) => ({ [`#${attrName}`]: attrName }))
-    .reduce(merge, {});
-
-  const ExpressionAttributeValues = keys
-    .filter((attrName) => Item[attrName] !== undefined && Item[attrName] !== null)
-    .map((attrName) => ({ [`:${attrName}`]: Item[attrName] }))
-    .reduce(merge, {});
-
-  let UpdateExpression = `SET ${keys
-    .filter((attrName) => Item[attrName] !== undefined && Item[attrName] !== null)
-    .map((attrName) => `#${attrName} = :${attrName}`)
-    .join(', ')}`;
-
-  const UpdateExpressionRemove = keys
-    .filter((attrName) => Item[attrName] === null)
-    .map((attrName) => `#${attrName}`)
-    .join(', ');
-
-  if (UpdateExpressionRemove.length) {
-    UpdateExpression = `${UpdateExpression} REMOVE ${UpdateExpressionRemove}`;
-  }
-
-  return {
-    ExpressionAttributeNames,
-    ExpressionAttributeValues,
-    UpdateExpression,
-    ReturnValues: 'ALL_NEW',
-  };
-};
-
-export const timestampCondition = (fieldName = 'timestamp') => ({
-  ConditionExpression: `attribute_not_exists(#${fieldName}) OR #${fieldName} < :${fieldName}`,
-});
-
-export const pkCondition = (fieldName = 'pk') => ({
-  ConditionExpression: `attribute_not_exists(${fieldName})`,
-});
-
-export const updateDynamoDB = ({
-  debug = d('dynamodb'),
-  tableName = process.env.ENTITY_TABLE_NAME || process.env.EVENT_TABLE_NAME,
-  updateRequestField = 'updateRequest',
-  updateResponseField = 'updateResponse',
-  parallel = Number(process.env.UPDATE_PARALLEL) || Number(process.env.PARALLEL) || 4,
-  timeout = Number(process.env.DYNAMODB_TIMEOUT) || Number(process.env.TIMEOUT) || 1000,
-  ...opt
-} = {}) => {
-  const connector = new Connector({ debug, tableName, timeout });
-
-  const invoke = (uow) => {
-    if (!uow[updateRequestField]) return _(Promise.resolve(uow));
-
-    const p = connector.update(uow[updateRequestField])
-      .then((updateResponse) => ({ ...uow, [updateResponseField]: updateResponse }))
-      .catch(rejectWithFault(uow));
-
-    return _(p); // wrap promise in a stream
-  };
-
-  return (s) => s
-    .through(ratelimit(opt))
-    .map(invoke)
-    .parallel(parallel);
-};
-
-export const update = updateDynamoDB; // deprecated export
-
-export const putDynamoDB = ({
-  debug = d('dynamodb'),
-  tableName = process.env.EVENT_TABLE_NAME || process.env.ENTITY_TABLE_NAME,
-  putRequestField = 'putRequest',
-  parallel = Number(process.env.UPDATE_PARALLEL) || Number(process.env.PARALLEL) || 4,
-  timeout = Number(process.env.DYNAMODB_TIMEOUT) || Number(process.env.TIMEOUT) || 1000,
-} = {}) => {
-  const connector = new Connector({ debug, tableName, timeout });
-
-  const invoke = (uow) => {
-    if (!uow[putRequestField]) return _(Promise.resolve(uow));
-
-    const p = connector.put(uow[putRequestField])
-      .then((putResponse) => ({ ...uow, putResponse }))
-      .catch(rejectWithFault(uow));
-
-    return _(p); // wrap promise in a stream
-  };
-
-  return (s) => s
-    .map(invoke)
-    .parallel(parallel);
-};
-
-export const put = putDynamoDB; // deprecated export
+import { rejectWithFault } from '../utils/faults';
+import { debug as d } from '../utils/print';
 
 export const batchGetDynamoDB = ({
   debug = d('dynamodb'),
@@ -120,17 +20,27 @@ export const batchGetDynamoDB = ({
   const invoke = (uow) => {
     if (!uow[batchGetRequestField]) return _(Promise.resolve(uow));
 
-    const p = connector.batchGet(uow[batchGetRequestField])
-      .then(async (batchGetResponse) => ({
-        ...batchGetResponse,
-        Responses: await Object.keys(batchGetResponse.Responses).reduce(async (a, c) => {
-          a = await a;
-          return {
-            ...a,
-            [c]: await Promise.all(batchGetResponse.Responses[c]?.map(decrypt)),
-          };
-        }, {}),
-      }))
+    const req = JSON.stringify(uow[batchGetRequestField]);
+    const cached = memoryCache.get(req);
+
+    const p = (cached
+      ? /* istanbul ignore next */ Promise.resolve(cached)
+      : connector.batchGet(uow[batchGetRequestField])
+        .then(async (batchGetResponse) => ({
+          ...batchGetResponse,
+          Responses: await Object.keys(batchGetResponse.Responses).reduce(async (a, c) => {
+            a = await a;
+            return {
+              ...a,
+              [c]: await Promise.all(batchGetResponse.Responses[c]?.map(decrypt)),
+            };
+          }, {}),
+        }))
+        .then((batchGetResponse) => {
+          memoryCache.put(req, batchGetResponse);
+          return batchGetResponse;
+        })
+    )
       .then((batchGetResponse) => ({ ...uow, [batchGetResponseField]: batchGetResponse }))
       .catch(rejectWithFault(uow));
 
@@ -141,8 +51,6 @@ export const batchGetDynamoDB = ({
     .map(invoke)
     .parallel(parallel);
 };
-
-export const batchGet = batchGetDynamoDB; // deprecated export
 
 export const queryAllDynamoDB = (/* istanbul ignore next */{
   debug = d('dynamodb'),
@@ -180,10 +88,6 @@ export const queryAllDynamoDB = (/* istanbul ignore next */{
     .map(invoke)
     .parallel(parallel);
 };
-
-export const query = queryAllDynamoDB; // deprecated export
-
-export const queryDynamoDB = queryAllDynamoDB; // deprecated export
 
 export const toPkQueryRequest = (uow, rule) => ({
   KeyConditionExpression: '#pk = :pk',
@@ -300,8 +204,6 @@ export const scanSplitDynamoDB = ({
     .map(scan)
     .parallel(parallel);
 };
-
-export const scanDynamoDB = scanSplitDynamoDB; // deprecated export
 
 export const querySplitDynamoDB = ({
   debug = d('dynamodb'),
