+++ src/queries/index.js
@@ -1,0 +1,3 @@
+export * from './dynamodb';
+export * from './secretsmgr';
+export * from './s3';
