+++ src/queries/secretsmgr.js
