+++ src/sinks/fetch.js
@@ -2,9 +2,9 @@ import _ from 'highland';
 
 import Connector from '../connectors/fetch';
 
-import { rejectWithFault } from './faults';
-import { debug as d } from './print';
-import { ratelimit } from './ratelimit';
+import { rejectWithFault } from '../utils/faults';
+import { debug as d } from '../utils/print';
+import { ratelimit } from '../utils/ratelimit';
 
 export const fetch = ({ // eslint-disable-line import/prefer-default-export
   debug = d('fetch'),
