+++ src/sinks/index.js
@@ -1,0 +1,9 @@
+export * from './cloudwatch';
+export * from './dynamodb';
+export * from './eventbridge';
+export * from './firehose';
+export * from './kinesis';
+export * from './lambda';
+export * from './s3';
+export * from './sns';
+export * from './sqs';
