+++ src/utils/index.js
@@ -1,25 +1,15 @@
 export * from './batch';
 export * from './claimcheck';
-export * from './cloudwatch';
 export * from './compression';
-export * from './dynamodb';
 export * from './encryption';
-export * from './eventbridge';
 export * from './faults';
-export * from './firehose';
 export * from './handler';
-export * from './kinesis';
-export * from './lambda';
 export * from './log';
 export * from './opt';
 export * from './print';
 export * from './ratelimit';
 export * from './retry';
-export * from './secretsmgr';
-export * from './s3';
-export * from './sns';
 export * from './split';
-export * from './sqs';
 export * from './tags';
 export * from './time';
 export * from './uow';
