+++ test/unit/flavors/update.test.js
@@ -13,11 +13,13 @@ import { toDynamodbRecords, fromDynamodb } from '../../../src/from/dynamodb';
 import { defaultOptions } from '../../../src/utils/opt';
 import {
   toPkQueryRequest, toGetRequest,
+} from '../../../src/queries/dynamodb';
+import {
   updateExpression, timestampCondition,
-} from '../../../src/utils/dynamodb';
+} from '../../../src/sinks/dynamodb';
 import { DynamoDBConnector } from '../../../src/connectors';
 
-import { upd as update } from '../../../src/flavors/update';
+import { update } from '../../../src/flavors/update';
 
 describe('flavors/update.js', () => {
   beforeEach(() => {
