+++ test/unit/sinks/cloudwatch.test.js
@@ -3,38 +3,38 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import _ from 'highland';
 
-import { invokeLambda } from '../../../src/utils/lambda';
+import { putMetrics } from '../../../src/sinks/cloudwatch';
 
-import Connector from '../../../src/connectors/lambda';
+import Connector from '../../../src/connectors/cloudwatch';
 
-describe('utils/lambda.js', () => {
+describe('utils/cloudwatch.js', () => {
   afterEach(sinon.restore);
 
-  it('should invoke lambda', (done) => {
-    const stub = sinon.stub(Connector.prototype, 'invoke').resolves({});
+  it('should putMetrics', (done) => {
+    const stub = sinon.stub(Connector.prototype, 'put').resolves({});
 
     const uows = [{
-      invokeRequest: {
-        FunctionName: 'helloworld',
+      putRequest: {
+        Namespace: 'ns',
       },
     }];
 
     _(uows)
-      .through(invokeLambda())
+      .through(putMetrics())
       .collect()
       .tap((collected) => {
         // console.log(JSON.stringify(collected, null, 2));
 
         expect(stub).to.have.been.calledWith({
-          FunctionName: 'helloworld',
+          Namespace: 'ns',
         });
 
         expect(collected.length).to.equal(1);
         expect(collected[0]).to.deep.equal({
-          invokeRequest: {
-            FunctionName: 'helloworld',
+          putRequest: {
+            Namespace: 'ns',
           },
-          invokeResponse: {},
+          putResponse: {},
         });
       })
       .done(done);
